+++ src/main/java/nextstep/subway/path/application/PathService.java
@@ -2,8 +2,6 @@ package nextstep.subway.path.application;
 
 import nextstep.subway.line.application.LineService;
 import nextstep.subway.path.domain.SubwayMap;
-import nextstep.subway.path.domain.Path;
-import nextstep.subway.path.domain.PathFinder;
 import nextstep.subway.path.dto.PathResponse;
 import nextstep.subway.station.application.StationService;
 import nextstep.subway.station.domain.Station;
@@ -27,7 +25,6 @@ public class PathService {
         Station source = stationService.findById(sourceId);
         Station target = stationService.findById(targetId);
 
-        Path path = new PathFinder(subwayMap).findShortestPath(source, target);
-        return PathResponse.of(path);
+        return PathResponse.of(subwayMap.findShortestPath(source, target));
     }
 }
